Setup planned sections
