+++ Demos/Blazorise.Demo/Layouts/MainLayout.razor
@@ -112,13 +112,125 @@
                         <BarItem>
                             <BarDropdown>
                                 <BarDropdownToggle>
+                                    <BarIcon IconName="IconName.CompactDisc" />
+                                    Datagrid
+                                </BarDropdownToggle>
+                                <BarDropdownMenu>
+                                <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Data
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+                                        <BarDropdownItem To="tests/datagrid/static">
+                                            In Memory
+                                        </BarDropdownItem>
+                                        <BarDropdownItem To="tests/datagrid/manual">
+                                            From an External Source
+                                        </BarDropdownItem>
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+                                <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Virtualize
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+                                <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Aggregates
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+                                <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Filter
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+                                <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Sort
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+
+                                <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Resize
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+
+                                 <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Fixed Header
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+
+                                <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Selection
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+                                        <BarDropdownItem To="tests/datagrid/single">
+                                            Single
+                                        </BarDropdownItem>
+                                        <BarDropdownItem To="tests/datagrid/multiple">
+                                            Multiple
+                                        </BarDropdownItem>
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+
+                                <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Row Detail
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+
+                                <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Editing
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+
+
+                                 <BarDropdown>
+                                    <BarDropdownToggle>
+                                        Pager
+                                    </BarDropdownToggle>
+                                    <BarDropdownMenu>
+
+                                    </BarDropdownMenu>
+                                </BarDropdown>
+                                </BarDropdownMenu>
+                            </BarDropdown>
+                        </BarItem>
+                        <BarItem>
+                            <BarDropdown>
+                                <BarDropdownToggle>
                                     <BarIcon IconName="IconName.PuzzlePiece" />
                                     Extensions
                                 </BarDropdownToggle>
                                 <BarDropdownMenu>
-                                    <BarDropdownItem To="tests/datagrid">
-                                        Data Grid
-                                    </BarDropdownItem>
                                     <BarDropdownItem To="tests/components">
                                         Components
                                     </BarDropdownItem>
@@ -237,16 +349,18 @@
                             <BarDropdown>
                                 <BarDropdownToggle><Icon Name="IconName.Language" /></BarDropdownToggle>
                                 <BarDropdownMenu RightAligned="true">
-                                    @foreach ( var cultureInfo in LocalizationService.AvailableCultures )
+                                    @foreach (var cultureInfo in LocalizationService.AvailableCultures)
                                     {
                                         <BarDropdownItem @key="@cultureInfo.Name" Clicked="@(()=>SelectCulture(cultureInfo.Name))">
-                                            @if ( cultureInfo.IsNeutralCulture )
+                                            @if (cultureInfo.IsNeutralCulture)
                                             {
-                                                @cultureInfo.EnglishName;
+                                                @cultureInfo.EnglishName
+                                                ;
                                             }
                                             else
                                             {
-                                                @cultureInfo.Parent.EnglishName;
+                                                @cultureInfo.Parent.EnglishName
+                                                ;
                                             }
                                         </BarDropdownItem>
                                     }
@@ -296,6 +410,6 @@
         </LayoutContent>
     </Layout>
 </Layout>
-@code{
+@code {
     RenderFragment customIcon =@<img src="/brand-logo.png" style="width:32px; height: 32px" />;
 }
